Avoid adding new unneeded names to blocks in text roundtripping  (#4943)
Previously the wat parser would turn this input:

(block
  (nop)
)

into something like this:

(block $block17
  (nop)
)

It just added a name all the time, in case the block is referred to by an index
later even though it doesn't have a name.

This PR makes us rountrip more precisely by not adding such names: if there
was no name before, and there is no break by index, then do not add a name.

In addition, this will be useful for non-nullable locals since whether a block has
a name or not matters there. Like #4912, this makes us more regular in our
usage of block names.