OptimizeInstructions: Select => and/or in more cases (#4154)
x ? 0 : y  ==>   z & y   where z = !x

x ? y : 1  ==>   z | y   where z = !x

Only do this when we have z = !x, that is, we can invert x without adding
an actual eqz (which would add work).

To do this, canonicalize selects to prefer to flip the arms, when
possible, if it would move a constant to a location that the existing
optimizations already turn into an and/or. That is,

x >= 5 ? 0 : y != 42

would be canonicalized into

x < 5 ? y != 42 : 0

and existing opts turn that into

(x < 5) & (y != 42)

The canonicalization does not always help this optimization, as we need
the values to be boolean to do this, but canonicalizing is still nice to get
more regular code which might compress slightly better.