Fix compilation errors with C++20
In C++20 makes operator calls become ambiguous wrt to a rewritten
version of itself with swapped candidates.

See https://gcc.godbolt.org/z/5q7GTr3bM for an example.

This commit fixes an occurence of this breakage in
`parent_index_iterator.h`. Note that the actual breakage is in
`passes/Poppify.cpp` that exposes the constraints of comparison
operator for `std::reverse_iteratot<Iterator>`.