OptimizeInstructions: Optimize comparisons with an added offset (#5025)
E.g.

x + C1 > C2  ==>  x > (C2-C1)

We do need to be careful of overflows in either the add on the left or
the proposed subtract on the right. In the latter case, we can at least do

x + C1 > C2  ==>  x + (C1-C2) > 0

Helps #5008 (but more patterns remain).

Found by the superoptimizer #4994. This was the top suggestion for Java and Dart.