[GUFA] Optimize functions not taken by reference better (#5085)
This moves the logic to add connections from signatures to functions from the top level
into the RefFunc logic. That way we only add those connections to functions that
actually have a RefFunc, which avoids us thinking that a function without one can be
reached by a call_ref of its type.

Has a small but non-zero benefit on j2wasm.