[Wasm GC] Use Cones in GUFA data reads and writes (#5157)
When we read from a struct/array using a cone type, read from the types in the cone
and nothing else. Previously we used the declared type in the wasm, which might be
larger (both in the base type and the depth). Likewise, in a write.

To do this, this extends ConeReadLocation with a depth (previously the depth there
was assumed to be infinite, and now it is to a potentially limited depth).

After this we are fully utilizing cone types in GUFA, as the test changes show (or at
least I can't think of any other uses of cones).