[NFC] Remove an ancient hack in ReFinalize (#5183)
Poking in the git history, this was some kind of hack for a problem that appears to
no longer exist since at least 5 years ago.

Logically, refinalize should never change a type from unreachable to none, or at
least if we have a place that does this, we should manually do the necessary things
around that, like updating the function's return type. The opposite, none (or anything
else) to unreachable is the common case, where we use refinalize to propagate
that type upwards.

Fuzzing also finds no issues.