Move removable code in CodePushing (#5187)
This is safe since we "partially remove" it: we don't move it to a place it might
execute more, but make it possibly execute less. See the new comment for
more details.

Motivated by wasm GC but this can help wasm MVP as well. In both cases
loads from memory can trap, which limits what the VM can do to optimize
them past conditions, but in trapsNeverHappens we can do that at the
toolchain level:

x = read();
if () { .. }
use(x);

=>

if () { .. }
x = read(); // moved to here, and might not execute if the if did a break/return
use(x);