CodePushing: Push into If arms (#5191)
Previously the pass only pushed past an if or a br_if. This does the same but into an
if arm. On Wasm GC for example this can perform allocation sinking:

function foo() {
  x = new A();
  if (..) {
    use(x);
  }
}

=>

function foo() {
  if (..) {
    x = new A(); // this moved
    use(x);
  }
}

The allocation won't happen if we never enter the if. This helps wasm MVP too,
and in fact some existing tests benefit.