RedundantSetElimination: Look at fallthrough values (#5213)
This can help in rare cases in MVP wasm, say for the return value of a block. But for
wasm GC it is very important due to casts.

Similar logic was added as part of #5194 for SimplifyLocals. It should probably have
been in a separate PR then. This does the right thing for RedundantSetElimination,
as a separate PR. Full tests will appear in that later PR (it is not really possible to test
the GC side yet - we need the logic in the later PR that actually switches to a more
refined local index when available).