Fix binary parsing of data segment memory (#5208)
The binary parser was eagerly getting the name of memories to set the `memory`
field of data segments, but that meant that when the memory names were updated
later while parsing the names section, the data segment memory fields would
become out of date. Update the issue by deferring setting the `memory` fields
like we do for other parts of IR that reference memories.

Also fix a segfault in the validator that was triggered by the reproducer for
this bug before the bug was fixed.

Fixes #5204.