[Wasm GC] RSE: Switch local.get to use a more refined type when possible (#5216)
Similar to #5194 but for RedundantSetElimination. This has similar benefits in terms
of using a more refined local in hopes of avoiding casts in followup opts, but unlike
SimplifyLocals this will operate across basic blocks.

To do this, we need to track not just local.set but also local.get in that pass. Then
in each basic block we can track the equivalent locals and pick from them.

I see a few dozen casts removed in the J2Wasm binary. Often stuff like this happens:

y = cast(x);
if (..) {
  foo(x); // this could use y
}